schema changes applied only to master
