+++ sirepo/package_data/static/js/sirepo.js
@@ -699,6 +699,7 @@ SIREPO.app.factory('notificationService', function(cookieService, $sce) {
 SIREPO.app.service('validationService', function(utilities) {
 
     this.fieldValidators = {};
+    this.enumValidators = {};
 
     this.setFieldValidator = function(name, validatorFn, messageFn) {
         if(! this.fieldValidators[name]) {
@@ -706,6 +707,7 @@ SIREPO.app.service('validationService', function(utilities) {
         }
         this.fieldValidators[name].vFunc = validatorFn;
         this.fieldValidators[name].vMsg = messageFn;
+        return this.fieldValidators[name];
     };
     this.getFieldValidator = function(name) {
         return this.fieldValidators[name];
@@ -731,6 +733,35 @@ SIREPO.app.service('validationService', function(utilities) {
         return fv ? (! ngModel.$valid ? fv.vMsg(ngModel.$viewValue) : '') : '';
     };
 
+    // lazy creation of validator, plus special handling
+    this.getEnumValidator = function(enumName) {
+
+        var validator = this.getFieldValidator(enumName);
+        if(validator) {
+            return validator;
+        }
+        var enums = SIREPO.APP_SCHEMA.enum[enumName];
+        if(! enums) {
+            throw enumName + ':' + ' no such enum';
+        }
+        var isValid = function(name) {
+            return enums.map(function (e) {
+                return e[SIREPO.ENUM_INDEX_VALUE];
+            }).indexOf(name) >= 0;
+        };
+        var err = function(name) {
+            return name + ':' + ' no such value in ' + enumName;
+        };
+        validator = this.setFieldValidator(enumName, isValid, err);
+        validator.find = function (name) {
+            if(! validator.vFunc(name)) {
+                throw validator.vMsg(name);
+            }
+            return name;
+        };
+        return validator;
+    };
+
     this.validateFieldOfType = function(value, type) {
         if (value === undefined || value === null || value === '')  {
             // null files OK, at least sometimes
@@ -751,6 +782,9 @@ SIREPO.app.service('validationService', function(utilities) {
         }
         if (type === 'String') {
             return true;
+        }
+        if(SIREPO.APP_SCHEMA.enum[type]) {
+            return this.getEnumValidator(type).vFunc(value);
         }
         // TODO(mvk): other types here, for now just accept everything
         return true;
@@ -889,7 +923,7 @@ SIREPO.app.factory('frameCache', function(appState, panelState, requestSender, $
     return self;
 });
 
-SIREPO.app.factory('panelState', function(appState, requestSender, simulationQueue, $compile, $rootScope, $timeout, $window) {
+SIREPO.app.factory('panelState', function(appState, requestSender, simulationQueue, validationService, $compile, $rootScope, $timeout, $window) {
     // Tracks the data, error, hidden and loading values
     var self = {};
     var panels = {};
