+++ sirepo/package_data/static/json/schema-common.json
@@ -105,6 +105,7 @@
         },
         "constants": {
             "clientCookie": "sirepo_cookie_js",
+            "maxSimCopies": 100,
             "oneDayMillis": 86400000
         },
         "cookies": {
